+++ samples/GiraffeRazorSample/Program.fs
@@ -8,6 +8,7 @@ open Microsoft.AspNetCore.Builder
 open Microsoft.AspNetCore.Hosting
 open Microsoft.AspNetCore.Http
 open Microsoft.AspNetCore.Http.Features
+open Microsoft.AspNetCore.Mvc.ModelBinding
 open Microsoft.Extensions.Logging
 open Microsoft.Extensions.DependencyInjection
 open FSharp.Control.Tasks.V2.ContextInsensitive
@@ -24,9 +25,21 @@ type Person =
         Name : string
     }
 
+[<CLIMutable>]
+type CreatePerson =
+    {
+        Name    : string
+        CheckMe : bool
+    }
+
 // ---------------------------------
 // Web app
 // ---------------------------------
+
+let antiforgeryTokenHandler =
+    text "Bad antiforgery token" 
+    |> RequestErrors.badRequest
+    |> validateAntiforgeryToken
 
 let bytesToKbStr (bytes : int64) =
     sprintf "%ikb" (bytes / 1024L)
@@ -55,9 +68,38 @@ let largeFileUploadHandler =
         }
 
 let renderPerson =
-    let model = { Name = "Razor" }
-    let viewData = dict [("Title", box "Mr Fox")]
-    razorHtmlView "Person" (Some model) (Some viewData)
+    fun (next : HttpFunc) (ctx : HttpContext) ->
+        task {
+            let name =
+                match ctx.TryGetQueryStringValue "name" with
+                | Some n -> n
+                | None -> "Razor"
+            let model = { Name = name }
+            let viewData = dict [("Title", box "Mr Fox")]
+            return! razorHtmlView "Person" (Some model) (Some viewData) None next ctx
+        }
+
+let renderCreatePerson =
+    let model = { Name = ""; CheckMe = true }
+    let viewData = dict [("Title", box "Create peson")]
+    razorHtmlView "CreatePerson" (Some model) (Some viewData) None
+
+let createPerson =
+    fun (next : HttpFunc) (ctx : HttpContext) ->
+        task {
+            let modelState = ModelStateDictionary()
+            let! model = ctx.BindModelAsync<CreatePerson>()
+            if not model.CheckMe then
+                modelState.AddModelError("CheckMe", "Checkbox must be checked")
+                modelState.AddModelError(String.Empty, "Error without an associated field")
+            if String.IsNullOrWhiteSpace(model.Name) then modelState.AddModelError("Name", "Name is rquired")
+            if modelState.IsValid then
+                let url = sprintf "/person?name=%s" model.Name
+                return! redirectTo false url next ctx
+            else
+                let viewData = dict [("Title", box "Create peson")]
+                return! razorHtmlView "CreatePerson" (Some model) (Some viewData) (Some modelState) next ctx
+        }
 
 let viewData =
     dict [
@@ -70,15 +112,17 @@ let webApp =
     choose [
         GET >=>
             choose [
-                route  "/"       >=> text "index"
-                route  "/razor"  >=> razorView "text/html" "Hello" None (Some viewData)
-                route  "/person" >=> renderPerson
-                route  "/upload" >=> razorHtmlView "FileUpload" (Some "File upload") (Some viewData)
+                route  "/"              >=> text "index"
+                route  "/razor"         >=> razorView "text/html" "Hello" None (Some viewData) None
+                route  "/person/create" >=> renderCreatePerson
+                route  "/person"        >=> renderPerson
+                route  "/upload"        >=> razorHtmlView "FileUpload" (Some "File upload") (Some viewData) None
             ]
         POST >=>
             choose [
-                route "/small-upload" >=> smallFileUploadHandler
-                route "/large-upload" >=> largeFileUploadHandler
+                route "/small-upload"  >=> smallFileUploadHandler
+                route "/large-upload"  >=> largeFileUploadHandler
+                route "/person/create" >=> antiforgeryTokenHandler >=> createPerson
             ]
         text "Not Found" |> RequestErrors.notFound ]